Rename files to fit HashiCorp guidelines
