+++ signalfx/resource_signalfx_detector.go
