+++ signalfx/resource_signalfx_single_value_chart.go
