+++ signalfx/resource_signalfx_text_chart.go
